split up solution into different files; #22
